requirements.txt
absl-py==2.1.0
art==6.2
cachetools==4.2.4
charset-normalizer==3.3.2
cycler==0.11.0
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.62.2
huggingface-hub==0.16.4
idna==3.7
imageio==2.9.0
importlib-metadata==6.7.0
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.5.3.56
packaging==24.0
pandas==0.25.3
parameterized==0.9.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycm==3.2
pydicom==2.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.3.2
tokenizers==0.13.3
tqdm==4.62.1
transformers==4.30.2
typing_extensions==4.7.1
uncertainty-calibration==0.1.4
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0